restrict AllowedIPs at client side
Restrict the AllowedIPs at client side to the tunnel subnet so that only
the traffic targeting the tunnel subnet will be routed to the server.
This matches the common usage of VPN in NAS scenario that the client may
setup a channel to server to access its services while let other network
traffic go through the default gateway on the client.